requirements.txt
psycopg2-binary==2.9.5
pandas==1.5.3
requests==2.28.2
python-dotenv==0.21.1